bf16

+++ core/model.py
@@ -40,7 +40,8 @@ def load_model(model_name: str, tokenizer: PreTrainedTokenizerFast, context_leng
         pretrained_name = model_name
 
     model = LlamaForCausalLM.from_pretrained(pretrained_name, config=model_config, token=hf_api_token,
-                                                state_dict=state_dict,)
+                                             torch_dtype=torch.bfloat16,
+                                                state_dict=state_dict)
     model.config.pad_token_id = tokenizer.pad_token_id
 
     apply_activation_checkpointing(model)
